+++ IMU/main/Mahony_Filter.h
@@ -21,12 +21,12 @@ class Mahony_Filter : public Fusion_Filter
 {
 private:
     IMU::Quaternion quaternion;
-    const float MAHONY_KP, MAHONY_KI;
+    const double MAHONY_KP, MAHONY_KI;
     timeval last_time;
     IMU::Sensor_Data error_integral;
 
 public:
-    Mahony_Filter(float kp, float ki);
+    Mahony_Filter(double kp, double ki);
     ~Mahony_Filter() override = default;
 
     void Compute(IMU::Sensor_Data accel, IMU::Sensor_Data gyro, IMU::Euler_Angles &euler) override;
